requirements.txt
glfw==2.10.0
imgui==2.0.0
numpy==2.4.2
pyglm==2.8.3
PyOpenGL==3.1.10
